Apply HTML, CSS, and JavaScript best practices
