+++ html/about.html
@@ -5,8 +5,8 @@
   <meta http-equiv="X-UA-Compatible" content="IE=edge">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <script src="https://kit.fontawesome.com/2b1476549e.js" crossorigin="anonymous"></script>
-  <link rel="stylesheet" href="assets/stylesheets/mobile.css" media="screen and (max-width: 768px)">
-  <link rel="stylesheet" href="assets/stylesheets/desktop.css" media="screen and (min-width: 768px)">
+  <link rel="stylesheet" href="../stylesheets/mobile.css" media="screen and (max-width: 768px)">
+  <link rel="stylesheet" href="../stylesheets/desktop.css" media="screen and (min-width: 768px)">
   <link href="http://fonts.cdnfonts.com/css/cocogoose" rel="stylesheet">
   <link href="http://fonts.cdnfonts.com/css/lato" rel="stylesheet">             
   <title> Web Development Course - About Page</title>
@@ -37,7 +37,7 @@
       </ul>
     </div>
     <nav class="desktop-header-nav">
-      <a href="index.html"><img src="./assets/images/weblogo.png" alt="GST_Logo"></a>
+      <a href="index.html"><img src="../assets/images/weblogo.png" alt="GST_Logo"></a>
       <ul>
         <li><a href="#">About</a></li>
         <li><a href="#">Courses</a></li>
@@ -58,30 +58,29 @@
         their students. We only selected an expert per field so that
         each section of our curriculum offers mastery learning.
       </p>
-
-    <div class="contact-us">
-      <p>
-        Please contact us per Email for any further 
-        questions about Global School of Tech.
-      </p>
-      <a>globaltech@gst.org</a>
-    </div>
+      <div class="contact-us">
+        <p>
+          Please contact us per Email for any further 
+          questions about Global School of Tech.
+        </p>
+        <a>globaltech@gst.org</a>
+      </div>
     </section>
-
     <section class="about-logo">
       <div class="logo-descript">
         <h1>Global School of Tech 2021 Logo</h1>
         <hr>
-        <p>The Global School of Tech 2021 Logo is the same 
+        <p>
+          The Global School of Tech 2021 Logo is the same 
           logo that the organization has been using ever 
           since 2017. The logo was obtained from <a href="https://www.transparentpng.com/">www.transparentpng.com</a>
-          .</p>
+        </p>
       </div>
-      <div class="logo"> <img src="./assets/images/weblogo.png" alt="GST_Logo"></div>
+      <div class="logo">
+        <img src="../assets/images/weblogo.png" alt="GST_Logo">
+      </div>
     </section>
-
     <hr>
-
     <section class="testimonies">
       <p class="testimonies-title">
         See the GST graduates testimonies.
@@ -89,44 +88,44 @@
       <hr>
       <p class="testimonies-intro">
         Take a look at what our students from the past two years who have already landed 
-       have to say about Global School of Tech.
+         have to say about Global School of Tech.
       </p>
       <div class="testimonies-cards">
         <div class="card1">
-          <img src="assets/images/2020.jpg" alt="2020 statistics">
+          <img src="../assets/images/2020.jpg" alt="2020 statistics">
           <h1 class="year">2020</h1>
           <p class="statistics">96% of our students got employed</p>
         </div>
         <div class="card2">
-          <img src="assets/images/2019.jpg" alt="2019 statistics">
+          <img src="../assets/images/2019.jpg" alt="2019 statistics">
           <h1 class="year">2019</h1>
           <p class="statistics">89% of our students got employed</p>
         </div>
-      </div>
+       </div>
     </section>
-    </main>
+  </main>
 
-    <footer class=" about-footer desktop-footer">
-      <div class="partner desktop-partner">
-        <p>Partner</p>
-        <hr>
-        <ul>
-          <li><img src="assets/images/fortiss.png" alt="fortiss_logo"></li>
-          <li><img src="assets/images/it-daily.png" alt="it-daily_logo"></li>
-          <li><img src="assets/images/ix.png" alt="ix_logo"></li>
-          <li><img src="assets/images/tu.png" alt="tu_logo"></li>
-          <li><img src="assets/images/partner1.png" alt="dpunkt_logo"></li>
-        </ul>
-      </div>
-      <div>
-        <img src="./assets/images/weblogo.png" alt="GST_Logo" class="footer-logo">
-        <p>
-          2017 Global School of Tech.
-          Some rights reserved.
-        </p>
-      </div>
-    </footer>
+  <footer class=" about-footer desktop-footer">
+    <div class="partner desktop-partner">
+      <p>Partner</p>
+      <hr>
+      <ul>
+        <li><img src="../assets/images/fortiss.png" alt="fortiss_logo"></li>
+        <li><img src="../assets/images/it-daily.png" alt="it-daily_logo"></li>
+        <li><img src="../assets/images/ix.png" alt="ix_logo"></li>
+        <li><img src="../assets/images/tu.png" alt="tu_logo"></li>
+        <li><img src="../assets/images/partner1.png" alt="dpunkt_logo"></li>
+      </ul>
+    </div>
+    <div>
+      <img src="../assets/images/weblogo.png" alt="GST_Logo" class="footer-logo">
+      <p>
+        2017 Global School of Tech.
+        Some rights reserved.
+      </p>
+    </div>
+  </footer>
 
-    <script src="./assets/js/menu.js"></script>
+  <script type="module" src="../js/modules/menu.js"></script>
 </body>
 </html>
