+++ js/app-data/tutors.js
@@ -1,0 +1,52 @@
+const tutors = [
+  {
+    img: '../assets/images/Micheal.png',
+    alt: 'Tutor: Michael Pearson',
+    tutorName: 'Michael Pearson',
+    skill: 'Front-End Accessibility Expert',
+    description: `Has extensive experience with crafting font-end
+  technologies that support accessibility 
+  requirements and standards.`,
+  },
+  {
+    img: '../assets/images/Lindsey.png',
+    alt: 'Tutor: Lindsey Holie',
+    tutorName: 'Lindsey Holie',
+    skill: 'Front-End Engineer',
+    description: `Experience in computer science and using this skill in
+  front-end development.`,
+  },
+  {
+    img: '../assets/images/Amanda.png',
+    alt: 'Tutor: Amanda Simons',
+    tutorName: 'Amanda Simons',
+    skill: 'Front-End Dev Ops',
+    description: `Extensive experience with software development pertaining
+  to collaborations, integration, automation, and quality.`,
+  },
+  {
+    img: '../assets/images/Cody.png',
+    alt: 'Tutor: Cody James',
+    tutorName: 'Cody James',
+    skill: 'U.I Developer',
+    description: 'Highly skilled in interaction design skills.',
+  },
+  {
+    img: '../assets/images/John.png',
+    alt: 'Tutor: John Mahn',
+    tutorName: 'John Mahn',
+    skill: 'Back-End Developer',
+    description: `Experience in computer science and using this skill in
+  front-end development.`,
+  },
+  {
+    img: '../assets/images/Hannah.png',
+    alt: 'Tutor: Hannah Eugne',
+    tutorName: 'Hanna Eugne',
+    skill: 'Front-End Engineer',
+    description: `Experience in computer science and using this skill in
+  front-end development.`,
+  },
+];
+
+export default tutors;
