+++ js/app.js
@@ -1,0 +1,5 @@
+import menu from './modules/menu.js';
+import loadTutors from './modules/display-tutors.js';
+
+loadTutors();
+menu();
