+++ stylesheets/mobile.css
@@ -1,5 +1,3 @@
-/* Home Page */
-
 * {
   box-sizing: border-box;
 }
@@ -86,7 +84,12 @@ header {
 }
 
 .intro {
-  background-image: linear-gradient(rgba(211, 211, 211, 0.9), rgba(211, 211, 211, 0.9)), url(../images/webmobile.jpg);
+  background-image:
+    linear-gradient(
+      rgba(211, 211, 211, 0.9),
+      rgba(211, 211, 211, 0.9)
+    ),
+    url(../assets/images/webmobile.jpg);
   background-repeat: no-repeat;
   background-size: cover;
   height: 100%;
@@ -114,17 +117,17 @@ header {
   font-family: 'Lato', sans-serif;
 }
 
-/******************* Main Program ********************/
-
 .main-program {
-  background-image: url(../images/main-bg.png), linear-gradient(rgb(39, 42, 49), rgb(39, 42, 49));
+  background-image:
+    url(../assets/images/main-bg.png),
+    linear-gradient(rgb(39, 42, 49), rgb(39, 42, 49));
   background-repeat: repeat;
   background-size: 30px;
   border: solid;
   padding-bottom: 30px;
 }
 
-.main-program > p {
+.main-program > h1 {
   text-align: center;
   font-size: 17px;
   font-weight: bold;
@@ -187,8 +190,6 @@ footer hr,
   width: clamp(10rem, 80vw + 1rem, 30rem);
 }
 
-/******************* Tutors *****************/
-
 .tutors {
   padding: 0 2rem;
 }
@@ -221,7 +222,7 @@ footer hr,
 }
 
 .image-container {
-  background-image: url('../images/chess.jpg');
+  background-image: url('../assets/images/chess.jpg');
   background-size: 40% 40%;
   background-repeat: no-repeat;
   background-position-x: 0;
